@@ -4,23 +4,22 @@ import { HttpClient } from '@angular/common/http';
 @Component({
   selector: 'app-value',
   templateUrl: './value.component.html',
-  styleUrls: ['./value.component.css']
+  styleUrls: ['./value.component.css'],
 })
 export class ValueComponent implements OnInit {
   values: any;
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) {}
 
   ngOnInit() {
-    this.getvalues();
+    this.getValues();
   }
 
-  getvalues() {
+  getValues() {
     this.http.get('http://localhost:5000/api/values').subscribe(response => {
       this.values = response;
     }, error => {
       console.log(error);
     });
   }
-
 }
